+++ test/src/HoistedNodeModuleTest.ts
@@ -1,8 +1,7 @@
 import { assertPack, linuxDirTarget, verifyAsarFileTree, modifyPackageJson } from "./helpers/packTester"
-import { Platform } from "electron-builder"
+import { Platform, Arch, DIR_TARGET } from "electron-builder"
 import { outputFile } from "fs-extra"
 import * as path from "path"
-import { readAsarJson } from "app-builder-lib/out/asar/asar"
 
 test.ifAll("yarn workspace", () =>
   assertPack(
@@ -90,15 +89,189 @@ test.ifAll("pnpm es5-ext without hoisted config", () =>
           outputFile(path.join(projectDir, "pnpm-lock.yaml"), ""),
         ])
       },
-      packed: async context => {
-        expect(await readAsarJson(path.join(context.getResources(Platform.LINUX), "app.asar"), "node_modules/d/package.json")).toMatchSnapshot()
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
+    }
+  )
+)
+
+test.ifAll("pnpm optional dependencies", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: linuxDirTarget,
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "electron-clear-data": "^1.0.5",
+            }
+            data.optionalDependencies = {
+              debug: "3.1.0",
+            }
+          }),
+          outputFile(path.join(projectDir, "pnpm-lock.yaml"), ""),
+        ])
       },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
+    }
+  )
+)
+
+test.ifAll("yarn electron-clear-data", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: Platform.WINDOWS.createTarget(DIR_TARGET, Arch.x64),
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "electron-clear-data": "^1.0.5",
+            }
+            data.optionalDependencies = {
+              debug: "3.1.0",
+            }
+          }),
+          outputFile(path.join(projectDir, "yarn.lock"), ""),
+        ])
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.WINDOWS)),
+    }
+  )
+)
+
+test.ifAll("npm electron-clear-data", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: Platform.WINDOWS.createTarget(DIR_TARGET, Arch.x64),
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "electron-clear-data": "^1.0.5",
+            }
+            data.optionalDependencies = {
+              debug: "3.1.0",
+            }
+          }),
+          outputFile(path.join(projectDir, "package-lock.json"), ""),
+        ])
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.WINDOWS)),
+    }
+  )
+)
+
+// https://github.com/electron-userland/electron-builder/issues/8842
+test.ifAll("yarn some module add by manual instead of install", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: Platform.WINDOWS.createTarget(DIR_TARGET, Arch.x64),
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: async (projectDir, tmpDir) => {
+        await outputFile(path.join(projectDir, "yarn.lock"), "")
+        await outputFile(path.join(projectDir, "node_modules", "foo", "package.json"), `{"name":"foo","version":"9.0.0","main":"index.js","license":"MIT"}`)
+        await modifyPackageJson(projectDir, data => {
+          data.dependencies = {
+            debug: "3.1.0",
+          }
+        })
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.WINDOWS)),
+    }
+  )
+)
+
+//https://github.com/electron-userland/electron-builder/issues/8857
+test.ifAll("yarn max stack", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: linuxDirTarget,
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "npm-run-all": "^4.1.5",
+            }
+          }),
+          outputFile(path.join(projectDir, "yarn.lock"), ""),
+        ])
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
+    }
+  )
+)
+
+test.ifAll("pnpm max stack", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: linuxDirTarget,
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "npm-run-all": "^4.1.5",
+            }
+          }),
+          outputFile(path.join(projectDir, "pnpm-lock.yaml"), ""),
+        ])
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
+    }
+  )
+)
+
+//github.com/electron-userland/electron-builder/issues/8842
+test.ifAll("yarn ms", () =>
+  assertPack(
+    "test-app-hoisted",
+    {
+      targets: linuxDirTarget,
+    },
+    {
+      isInstallDepsBefore: true,
+      projectDirCreated: projectDir => {
+        return Promise.all([
+          modifyPackageJson(projectDir, data => {
+            data.dependencies = {
+              "@sentry/electron": "5.11.0",
+              "electron-clear-data": "^1.0.5",
+            }
+            data.devDependencies = {
+              electron: "34.0.2",
+            }
+          }),
+          outputFile(path.join(projectDir, "yarn.lock"), ""),
+        ])
+      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
     }
   )
 )
 
 //github.com/electron-userland/electron-builder/issues/8426
-https: test.ifAll("yarn parse-asn1", () =>
+test.ifAll("yarn parse-asn1", () =>
   assertPack(
     "test-app-hoisted",
     {
@@ -116,15 +289,13 @@ https: test.ifAll("yarn parse-asn1", () =>
           outputFile(path.join(projectDir, "yarn.lock"), ""),
         ])
       },
-      packed: async context => {
-        expect(await readAsarJson(path.join(context.getResources(Platform.LINUX), "app.asar"), "node_modules/asn1.js/package.json")).toMatchSnapshot()
-      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
     }
   )
 )
 
 //github.com/electron-userland/electron-builder/issues/8431
-https: test.ifAll("npm tar", () =>
+test.ifAll("npm tar", () =>
   assertPack(
     "test-app-hoisted",
     {
@@ -142,14 +313,7 @@ https: test.ifAll("npm tar", () =>
           outputFile(path.join(projectDir, "package-lock.json"), ""),
         ])
       },
-      packed: async context => {
-        let tar = await readAsarJson(path.join(context.getResources(Platform.LINUX), "app.asar"), "node_modules/tar/package.json")
-        let minipass = await readAsarJson(path.join(context.getResources(Platform.LINUX), "app.asar"), "node_modules/minipass/package.json")
-        let minizlib = await readAsarJson(path.join(context.getResources(Platform.LINUX), "app.asar"), "node_modules/minizlib/package.json")
-        expect(tar.version).toEqual("7.4.3")
-        expect(minipass.version).toEqual("7.1.2")
-        expect(minizlib.version).toEqual("3.0.1")
-      },
+      packed: context => verifyAsarFileTree(context.getResources(Platform.LINUX)),
     }
   )
 )